[ASTERIXDB-3268][STO] Add parallel downloader/cacher
- user model changes: no
- storage format changes: no
- interface changes: yes

Details:
Download uncached files in parallel. Also, when listing
files, we will ignore consulting with the cloud store
for downloaded indexes.

Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/17791
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ asterixdb/asterix-cloud/src/main/java/org/apache/asterix/cloud/bulk/DeleteBulkCloudOperation.java
@@ -50,13 +50,17 @@ public class DeleteBulkCloudOperation extends DeleteBulkOperation {
          *      Actually, is there a case where we delete multiple directories from the cloud?
          */
         List<String> paths = fileReferences.stream().map(FileReference::getRelativePath).collect(Collectors.toList());
+        if (paths.isEmpty()) {
+            return 0;
+        }
+
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Bulk deleting: local: {}, cloud: {}", fileReferences, paths);
         }
         cloudClient.deleteObjects(bucket, paths);
         // Bulk delete locally as well
-        int localDeletes = super.performOperation();
-        callBack.call(localDeletes, paths);
-        return paths.size();
+        super.performOperation();
+        callBack.call(fileReferences);
+        return fileReferences.size();
     }
 }
